Split tests into unit & integration test
